pyproject.toml
[tool.poetry]
name = "bessavagner-page"
version = "0.1.0"
description = ""
authors = ["bessavagner <bessavagner@gmail.com>"]
packages = [{include = "app", from = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohttp = "^3.11.11"
aiohttp-jinja2 = "^1.6"
python-dotenv = "^1.0.1"
watchfiles = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
